% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat2.r
\name{splat.list}
\alias{splat.list}
\title{splat list}
\usage{
splat.list(x, grid, proj = NULL, ...)
}
\arguments{
\item{x}{fit from SGAT}

\item{grid}{raster}

\item{proj}{crs}

\item{...}{ignored}
}
\description{
splat list
}

