% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-package.r, R/splat2.r
\docType{package}
\name{splat}
\alias{print.splat}
\alias{show.splat}
\alias{splat}
\alias{splat-package}
\title{splat.}
\usage{
splat(x, ...)

\method{print}{splat}(x, ...)

show.splat(x, ...)
}
\arguments{
\item{x}{generic}

\item{...}{ignored}
}
\description{
splat.

splat
}

